Rename animator modules to spirit_animator and drop Animator3D alias
SpiritAnimator lives in spirit_animator.py; SkeletalAnimator stays in skeletal_animator.py. Remove animator3d.py and public Animator3D export.

Made-with: Cursor
