+++ python/Infernux/components/__init__.py
@@ -81,9 +81,8 @@ from .decorators import (
     HelpURL,
     Icon,
 )
-from .animator2d import SpiritAnimator
+from .spirit_animator import SpiritAnimator
 from .skeletal_animator import SkeletalAnimator
-from .animator3d import Animator3D  # deprecated alias
 
 __all__ = [
     "InxComponent",
@@ -146,5 +145,4 @@ __all__ = [
     # Animation
     "SpiritAnimator",
     "SkeletalAnimator",
-    "Animator3D",
 ]
