--- python/Infernux/components/animator3d.py
@@ -1,3 +1,0 @@
-"""Backward compatibility: ``Animator3D`` was renamed to :class:`SkeletalAnimator`."""
-
-from .skeletal_animator import SkeletalAnimator as Animator3D  # noqa: F401
