+++ python/Infernux/components/spirit_animator.py
